requirements/base.txt
Django==1.8a1
bpython==0.13.2
PyYAML==3.11
dj-database-url==0.3.0
psycopg2==2.5.4
Pillow==2.7.0
djangorestframework==3.0.4
django-uuslug==1.0.3
django-reversion==1.8.5
Markdown==2.6
